renamed to show and moved to pipelines
