--- acceptance/experimental/preview-datasets/basic/script
@@ -1,11 +1,0 @@
-title "table output\n"
-trace $CLI experimental preview-datasets main.default.people
-
-title "json output\n"
-trace $CLI experimental preview-datasets main.default.people --output json
-
-title "invalid table name\n"
-musterr trace $CLI experimental preview-datasets people
-
-title "wire SQL pins quoting and default limit\n"
-trace print_requests.py //api/2.0/sql/statements
