+++ acceptance/pipelines/show/basic/script
@@ -1,0 +1,11 @@
+title "table output\n"
+trace $CLI pipelines show main.default.people
+
+title "json output\n"
+trace $CLI pipelines show main.default.people --output json
+
+title "invalid table name\n"
+musterr trace $CLI pipelines show people
+
+title "wire SQL pins quoting and default limit\n"
+trace print_requests.py //api/2.0/sql/statements
