+++ acceptance/pipelines/show/help/output.txt
@@ -1,18 +1,16 @@
 
 === help
->>> [CLI] experimental preview-datasets --help
+>>> [CLI] pipelines show --help
 Preview a table's columns and sample rows on a SQL warehouse.
 
 TABLE is a fully-qualified name: catalog.schema.table (Unity Catalog) or
 schema.table (legacy Hive metastore).
 
-Experimental and may change in future versions.
-
 Usage:
-  databricks experimental preview-datasets TABLE [flags]
+  databricks pipelines show TABLE [flags]
 
 Flags:
-  -h, --help                  help for preview-datasets
+  -h, --help                  help for show
   -n, --limit int             Maximum number of rows to fetch. (default 100)
       --warehouse-id string   SQL warehouse to run the query on. Defaults to DATABRICKS_WAREHOUSE_ID or a workspace default.
 
@@ -21,3 +19,4 @@ Global Flags:
   -o, --output type      output type: text or json (default text)
   -p, --profile string   ~/.databrickscfg profile
   -t, --target string    bundle target to use (if applicable)
+      --var strings      set values for variables defined in project config. Example: --var="foo=bar"
