+++ acceptance/pipelines/show/help/script
@@ -1,4 +1,4 @@
 # Pin the command tree so any change to the command or its help shows up as a diff here.
 
 title "help"
-trace $CLI experimental preview-datasets --help
+trace $CLI pipelines show --help
