+++ cmd/pipelines/show_identifier.go
@@ -1,4 +1,4 @@
-package previewdatasetscmd
+package pipelines
 
 import (
 	"fmt"
