+++ cmd/pipelines/show_render.go
@@ -1,4 +1,4 @@
-package previewdatasetscmd
+package pipelines
 
 import (
 	"bytes"
